Keep migration safety within shared project boundary
