+++ apps/desktop/tsdown.config.mts
@@ -3,24 +3,24 @@
 // Layer: Desktop build config
 // Depends on: tsdown.
 
+import { createHash } from "node:crypto";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
 
 import { defineConfig } from "tsdown";
-import {
-  MIGRATION_RUNTIME_SOURCE_RELATIVE_PATH,
-  migrationRuntimeSourceDigest,
-} from "../../packages/shared/src/migrationSafety.ts";
 
 const sourcemapEnv = process.env.SYNARA_DESKTOP_SOURCEMAP?.trim().toLowerCase();
 const buildSourcemap = sourcemapEnv === "1" || sourcemapEnv === "true";
 const windowsUpdaterPublisher = process.env.AZURE_TRUSTED_SIGNING_SUBJECT_DN?.trim() ?? "";
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
 const migrationRuntimeSource = fs.readFileSync(
-  path.join(repoRoot, MIGRATION_RUNTIME_SOURCE_RELATIVE_PATH),
+  path.join(repoRoot, "apps/server/src/persistence/Migrations.ts"),
   "utf8",
 );
+const migrationRuntimeSourceDigest = createHash("sha256")
+  .update(migrationRuntimeSource, "utf8")
+  .digest("hex");
 
 const shared = {
   format: "cjs" as const,
@@ -39,9 +39,7 @@ export default defineConfig([
     external: ["original-fs"],
     define: {
       __SYNARA_WINDOWS_UPDATER_PUBLISHER__: JSON.stringify(windowsUpdaterPublisher),
-      __SYNARA_MIGRATION_RUNTIME_SOURCE_DIGEST__: JSON.stringify(
-        migrationRuntimeSourceDigest(migrationRuntimeSource),
-      ),
+      __SYNARA_MIGRATION_RUNTIME_SOURCE_DIGEST__: JSON.stringify(migrationRuntimeSourceDigest),
     },
     noExternal: (id) => id.startsWith("@synara/"),
   },
