+++ apps/server/tsdown.config.ts
@@ -3,23 +3,23 @@
 // Layer: Server build config
 // Depends on: tsdown.
 
+import { createHash } from "node:crypto";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
 
 import { defineConfig } from "tsdown";
-import {
-  MIGRATION_RUNTIME_SOURCE_RELATIVE_PATH,
-  migrationRuntimeSourceDigest,
-} from "../../packages/shared/src/migrationSafety.ts";
 
 const sourcemapEnv = process.env.SYNARA_SERVER_SOURCEMAP?.trim().toLowerCase();
 const buildSourcemap = sourcemapEnv === "1" || sourcemapEnv === "true";
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
 const migrationRuntimeSource = fs.readFileSync(
-  path.join(repoRoot, MIGRATION_RUNTIME_SOURCE_RELATIVE_PATH),
+  path.join(repoRoot, "apps/server/src/persistence/Migrations.ts"),
   "utf8",
 );
+const migrationRuntimeSourceDigest = createHash("sha256")
+  .update(migrationRuntimeSource, "utf8")
+  .digest("hex");
 
 export default defineConfig({
   entry: ["src/index.ts", "src/restoreMigrationBackup.ts"],
@@ -33,9 +33,7 @@ export default defineConfig({
   external: [/^bun:/u],
   sourcemap: buildSourcemap,
   define: {
-    __SYNARA_MIGRATION_RUNTIME_SOURCE_DIGEST__: JSON.stringify(
-      migrationRuntimeSourceDigest(migrationRuntimeSource),
-    ),
+    __SYNARA_MIGRATION_RUNTIME_SOURCE_DIGEST__: JSON.stringify(migrationRuntimeSourceDigest),
   },
   clean: true,
   noExternal: (id) => id.startsWith("@synara/"),
